Refactor rental applications UI into components
Extract UI, helpers and styles for the agent rental applications page into separate modules. Added AppDetailModal, RentalApplicationBadges and rentalApplicationHelpers (formatters, palette, status config and validation). Moved global styles into src/styles/agent/AgentRentalApplications.css and updated RentalApplications.jsx to import the new components/helpers and css (removed large inline CSS and duplicate component code). This reorganizes the code for reuse, readability and easier maintenance without changing behavior.
