+++ src/components/agent/rentalApplications/RentalApplicationBadges.jsx
@@ -1,0 +1,107 @@
+import { useEffect } from "react";
+import { C, APP_STATUS_CFG } from "./rentalApplicationHelpers";
+
+// ─── Toast ────────────────────────────────────────────────────────────────────
+export function Toast({ msg, type = "success", onDone }) {
+  useEffect(() => { const t = setTimeout(onDone, 3200); return () => clearTimeout(t); }, [onDone]);
+  return (
+    <div style={{
+      position: "fixed", bottom: 26, right: 26, zIndex: 9999,
+      background: C.dark,
+      color: type === "error" ? "#f09090" : "#90c8a8",
+      padding: "11px 18px", borderRadius: 12, fontSize: 13,
+      boxShadow: "0 10px 36px rgba(0,0,0,0.32)",
+      border: `1px solid ${type === "error" ? "rgba(240,128,128,0.15)" : "rgba(144,200,168,0.15)"}`,
+      fontFamily: "'DM Sans',sans-serif", animation: "ra-toast 0.2s ease",
+      display: "flex", alignItems: "center", gap: 8,
+    }}>
+      {type === "error" ? "⚠️" : "✅"} {msg}
+    </div>
+  );
+}
+
+// ─── Skeleton ─────────────────────────────────────────────────────────────────
+export function Skeleton({ rows = 4, h = 54 }) {
+  return (
+    <div style={{ padding: "14px 16px", display: "flex", flexDirection: "column", gap: 10 }}>
+      {Array.from({ length: rows }).map((_, i) => (
+        <div key={i} className="ra-skeleton" style={{ height: h, opacity: 1 - i * 0.15 }} />
+      ))}
+    </div>
+  );
+}
+
+// ─── EmptyState ───────────────────────────────────────────────────────────────
+export function EmptyState({ icon, title, sub }) {
+  return (
+    <div style={{ padding: "52px 20px", textAlign: "center", color: C.textMut }}>
+      <div style={{ fontSize: 40, marginBottom: 12 }}>{icon}</div>
+      <p style={{ fontSize: 14, fontWeight: 600, color: C.textSub, margin: "0 0 4px",
+        fontFamily: "'Cormorant Garamond',Georgia,serif" }}>{title}</p>
+      {sub && <p style={{ fontSize: 12.5, margin: 0 }}>{sub}</p>}
+    </div>
+  );
+}
+
+// ─── StatusPill ───────────────────────────────────────────────────────────────
+export function StatusPill({ status }) {
+  const s = APP_STATUS_CFG[status] || { color: "#64748b", bg: "#f1f5f9", dot: "#64748b" };
+  return (
+    <span style={{
+      background: s.bg, color: s.color,
+      padding: "3px 10px", borderRadius: 20,
+      fontSize: 10.5, fontWeight: 600,
+      display: "inline-flex", alignItems: "center", gap: 5,
+      whiteSpace: "nowrap",
+    }}>
+      <span style={{ width: 5, height: 5, borderRadius: "50%", background: s.dot, display: "inline-block" }} />
+      {status}
+    </span>
+  );
+}
+
+// ─── SectionHeader ────────────────────────────────────────────────────────────
+export function SectionHeader({ title, count, children }) {
+  return (
+    <div style={{
+      padding: "14px 20px", borderBottom: `1px solid ${C.border}`,
+      display: "flex", alignItems: "center", justifyContent: "space-between",
+      background: "#fdf9f4", flexWrap: "wrap", gap: 10,
+    }}>
+      <div style={{ display: "flex", alignItems: "center", gap: 10 }}>
+        <p style={{ margin: 0, fontSize: 15, fontWeight: 700, color: C.text,
+          fontFamily: "'Cormorant Garamond',Georgia,serif" }}>{title}</p>
+        {count != null && (
+          <span style={{ background: `${C.gold}22`, color: C.textSub,
+            padding: "2px 9px", borderRadius: 20, fontSize: 11, fontWeight: 600 }}>
+            {count}
+          </span>
+        )}
+      </div>
+      {children}
+    </div>
+  );
+}
+
+// ─── StatCard ─────────────────────────────────────────────────────────────────
+export function StatCard({ icon, label, value, accent = C.gold }) {
+  return (
+    <div className="ra-stat ra-card ra-btn" style={{
+      padding: "18px 20px", display: "flex", alignItems: "center", gap: 14,
+    }}>
+      <div style={{
+        width: 46, height: 46, borderRadius: 12, flexShrink: 0,
+        background: `${accent}18`, border: `1.5px solid ${accent}28`,
+        display: "flex", alignItems: "center", justifyContent: "center", fontSize: 20,
+      }}>{icon}</div>
+      <div style={{ flex: 1, minWidth: 0 }}>
+        <p style={{ margin: 0, fontSize: 10, fontWeight: 600, color: C.textMut,
+          textTransform: "uppercase", letterSpacing: "0.9px", marginBottom: 3 }}>{label}</p>
+        <p style={{ margin: 0, fontSize: 26, fontWeight: 700, color: C.text,
+          fontFamily: "'Cormorant Garamond',Georgia,serif", letterSpacing: "-0.5px", lineHeight: 1 }}>
+          {value ?? "—"}
+        </p>
+      </div>
+    </div>
+  );
+}
