+++ src/styles/agent/AgentRentalApplications.css
@@ -1,0 +1,90 @@
+@import url('https://fonts.googleapis.com/css2?family=Cormorant+Garamond:ital,wght@0,400;0,600;0,700;1,400&family=DM+Sans:wght@300;400;500;600&display=swap');
+
+.ra * { box-sizing: border-box; }
+.ra { font-family: 'DM Sans', system-ui, sans-serif; background: #f2ede4; min-height: 100vh; }
+
+.ra-card {
+  background: #faf7f2;
+  border: 1.5px solid #e8e2d6;
+  border-radius: 14px;
+  box-shadow: 0 2px 16px rgba(20,16,10,0.06);
+  overflow: hidden;
+}
+
+.ra-btn {
+  transition: all 0.17s ease;
+  cursor: pointer;
+  font-family: 'DM Sans', sans-serif;
+  border: none;
+}
+.ra-btn:hover { opacity: 0.85; transform: translateY(-1px); }
+.ra-btn:disabled { opacity: 0.45; cursor: not-allowed; transform: none; }
+
+@keyframes ra-fade-up  { from{opacity:0;transform:translateY(18px)} to{opacity:1;transform:translateY(0)} }
+@keyframes ra-pulse    { 0%,100%{opacity:.4} 50%{opacity:.85} }
+@keyframes ra-toast    { from{opacity:0;transform:translateY(10px)} to{opacity:1;transform:translateY(0)} }
+@keyframes ra-spin     { to{transform:rotate(360deg)} }
+@keyframes ra-modal-in { from{opacity:0;transform:translateY(14px) scale(0.97)} to{opacity:1;transform:translateY(0) scale(1)} }
+
+.ra-section  { animation: ra-fade-up 0.5s ease 0.15s both; }
+.ra-row:hover { background: #f5f0e8 !important; }
+
+.ra-skeleton {
+  background: #ede9df;
+  border-radius: 10px;
+  animation: ra-pulse 1.5s ease infinite;
+}
+
+.ra-stat { animation: ra-fade-up 0.4s ease both; }
+.ra-stat:nth-child(1) { animation-delay: 0.05s; }
+.ra-stat:nth-child(2) { animation-delay: 0.10s; }
+.ra-stat:nth-child(3) { animation-delay: 0.15s; }
+.ra-stat:nth-child(4) { animation-delay: 0.20s; }
+
+.ra-table { width: 100%; border-collapse: collapse; }
+.ra-table th {
+  text-align: left; font-size: 10px; font-weight: 600;
+  color: #b0a890; text-transform: uppercase; letter-spacing: 0.9px;
+  padding: 10px 16px; background: #fdf9f4;
+  border-bottom: 1px solid #e8e2d6; white-space: nowrap;
+}
+.ra-table td {
+  padding: 13px 16px; font-size: 13px; color: #1a1714;
+  border-bottom: 1px solid #e8e2d6; vertical-align: middle;
+}
+.ra-table tr:last-child td { border-bottom: none; }
+.ra-table-wrap { overflow-x: auto; }
+
+.ra-input {
+  padding: 9px 13px;
+  border: 1.5px solid #e0dbd0;
+  border-radius: 10px;
+  font-size: 13.5px;
+  font-family: 'DM Sans', sans-serif;
+  color: #1a1714;
+  background: #f5f0e8;
+  outline: none;
+  transition: border-color 0.17s, box-shadow 0.17s, background 0.17s;
+}
+.ra-input:focus {
+  border-color: #c9b87a;
+  background: #faf7f2;
+  box-shadow: 0 0 0 3px rgba(201,184,122,0.12);
+}
+.ra-input::placeholder { color: #b0a890; }
+
+.ra-textarea {
+  width: 100%; padding: 10px 13px;
+  border: 1.5px solid #e0dbd0; border-radius: 10px;
+  font-size: 13.5px; font-family: 'DM Sans', sans-serif;
+  color: #1a1714; background: #f5f0e8;
+  outline: none; resize: vertical;
+  transition: border-color 0.17s, box-shadow 0.17s, background 0.17s;
+}
+.ra-textarea:focus {
+  border-color: #c9b87a; background: #faf7f2;
+  box-shadow: 0 0 0 3px rgba(201,184,122,0.12);
+}
+.ra-textarea::placeholder { color: #b0a890; }
+
+.ra-char-count { font-size: 11px; color: #b0a890; text-align: right; margin-top: 4px; }
